requirements.txt
oci
cryptography
PyJWT
fdk
pycryptodome
